Be more lenient with unfixable paths (#35)
This changes the behaviour slightly, in that if a path is found to be
not in the decendent path (and non_decendants == fix), then just allow
it, rather than returning undefined and making it an unusable path.

I still would like to see some kind of warning indicating that some of
the paths were unfixable, but this works for now in removing the
annoying error message, and also makes sync a little more useful if you
change the CWD or something.